Rename since it's not about opening the shell anymore
