+++ app/src/lib/is-git-on-path.ts
